+++ lib/src/jsony_base.dart
@@ -1,5 +1,7 @@
 import 'dart:convert';
 
+import 'package:jsony/jsony.dart';
+
 
 
 
@@ -7,10 +9,10 @@ import 'dart:convert';
 
 class JsonyVar{
 
-  late final String name;
+  late final String jsonyVarName;
   bool _match = true;
   bool get isMatched{ return _match; }
-  JsonyVar(this.name);
+  JsonyVar(this.jsonyVarName);
 
 }
 
@@ -21,7 +23,7 @@ class Jsony extends JsonyVar{
   List<dynamic> _variables = [];
 
 
-  Jsony(super.name);
+  Jsony(super.jsonyVarName);
 
 
   Jsony call() => this;
@@ -41,24 +43,24 @@ class Jsony extends JsonyVar{
     _variables.forEach((element) {
       result.addAll(element.obj);
     });
-    return {name: result};
+    return {jsonyVarName: result};
   }
 
 
-  void update(dynamic root){
-    Map<String, dynamic> r = root[name];
+  void updateElements(dynamic root){
+    Map<String, dynamic> r = root[jsonyVarName];
     _variables.forEach((element) {
-      bool exist = r.containsKey(element.name);
+      bool exist = r.containsKey(element.jsonyVarName);
       try{
-        if( !exist || (exist && root[name][element.name] != null) ) {
-          element.update(root[name]);
+        if( !exist || (exist && root[jsonyVarName][element.jsonyVarName] != null) ) {
+          element.updateElements(root[jsonyVarName]);
           if(!exist){ throw("Variable not found "); }
         }
         !element.isMatched ?_match=false :null;
       }
       catch(e){
         _match = false;
-        print("${e.toString()} at \"$name\":{\"${element.name}\": ?}");
+        print("${e.toString()} at \"$jsonyVarName\":{\"${element.jsonyVarName}\": ?}");
       }
     });
 
@@ -68,7 +70,7 @@ class Jsony extends JsonyVar{
   void fromJson(String jsonString){
     const JsonDecoder decoder = JsonDecoder();
     final Map<String,  dynamic> jsonObj = decoder.convert(jsonString);
-    update(jsonObj);
+    updateElements(jsonObj);
   }
 }
 
@@ -79,14 +81,14 @@ class JsonyBool extends JsonyVar{
 
   late bool value;
 
-  JsonyBool(super.name, this.value);
+  JsonyBool(super.jsonyVarName, this.value);
   bool call() => value;
-  String get string{  return '"$name":$value';  }
-  Map<String, bool> get obj{ return {name:value}; }
-  void update(dynamic v){
+  String get string{  return '"$jsonyVarName":$value';  }
+  Map<String, bool> get obj{ return {jsonyVarName:value}; }
+  void updateElements(dynamic v){
     _match=true;
-    v[name]!=null
-        ?value = bool.parse(v[name].toString())
+    v[jsonyVarName]!=null
+        ?value = bool.parse(v[jsonyVarName].toString())
         :_match=false;
   }
 }
@@ -98,14 +100,14 @@ class JsonyString extends JsonyVar{
 
   late String value;
 
-  JsonyString(super.name, this.value);
+  JsonyString(super.jsonyVarName, [String? value]) : value = value??"";
   String call() => value;
-  String get string{  return '"$name":"$value"';  }
-  Map<String, String> get obj{ return {name:value}; }
-  void update(dynamic v){
+  String get string{  return '"$jsonyVarName":"$value"';  }
+  Map<String, String> get obj{ return {jsonyVarName:value}; }
+  void updateElements(dynamic v){
     _match=true;
-    v[name]!=null
-      ?value = v[name].toString()
+    v[jsonyVarName]!=null
+      ?value = v[jsonyVarName].toString()
       :_match=false;
   }
 }
@@ -117,14 +119,14 @@ class JsonyInt extends JsonyVar{
 
   late int value;
 
-  JsonyInt(super.name, this.value);
+  JsonyInt(super.jsonyVarName, this.value);
   int call() => value;
-  String get string{  return '"$name":$value';  }
-  Map<String, int> get obj{ return {name:value}; }
-  void update(dynamic v){
+  String get string{  return '"$jsonyVarName":$value';  }
+  Map<String, int> get obj{ return {jsonyVarName:value}; }
+  void updateElements(dynamic v){
     _match=true;
-    v[name]!=null
-      ?value = int.parse(v[name].toString())
+    v[jsonyVarName]!=null
+      ?value = int.parse(v[jsonyVarName].toString())
       :_match=false;
   }
 }
@@ -132,17 +134,17 @@ class JsonyInt extends JsonyVar{
 
 
 
-class JsonyStringList extends JsonyVar{
+class JsonyListString extends JsonyVar{
 
   late List<String> value;
 
-  JsonyStringList(super.name, this.value);
+  JsonyListString(super.jsonyVarName, [List<String>? value]): value = value??[];
   List<String> call() => value;
-  Map<String,List<String>> get obj{ return{name: value}; }
-  void update(dynamic v){
+  Map<String,List<String>> get obj{ return{jsonyVarName: value}; }
+  void updateElements(dynamic v){
     _match=true;
-    v[name]!=null
-      ?value = List<String>.from(v[name] as List)
+    v[jsonyVarName]!=null
+      ?value = List<String>.from(v[jsonyVarName] as List)
       :_match=false;
   }
 }
@@ -150,17 +152,84 @@ class JsonyStringList extends JsonyVar{
 
 
 
-class JsonyDynamicList extends JsonyVar{
+class JsonyListDynamic extends JsonyVar{
 
   late List<dynamic> value;
 
-  JsonyDynamicList(super.name, this.value);
+  JsonyListDynamic(super.jsonyVarName, [List<dynamic>? value]): value = value??[];
   List<dynamic> call() => value;
-  Map<String,List<dynamic>> get obj{ return{name: value}; }
-  void update(dynamic v){
+  Map<String,List<dynamic>> get obj{ return{jsonyVarName: value}; }
+  void updateElements(dynamic v){
     _match=true;
-    v[name]!=null
-      ?value = List<dynamic>.from(v[name] as List)
+    v[jsonyVarName]!=null
+      ?value = List<dynamic>.from(v[jsonyVarName] as List)
       :_match=false;
+  }
+}
+
+
+
+
+class JsonyListListString extends JsonyVar{
+
+  late List<List<String>> value;
+
+  JsonyListListString(super.jsonyVarName, [List<List<String>>? value]): value = value??[];
+  List<List<String>> call() => value;
+  Map<String, List<List<String>>> get obj{ return{jsonyVarName: value}; }
+  void updateElements(dynamic v){
+    _match=true;
+    if( v[jsonyVarName]!=null ) {
+      value = [];
+      List<dynamic> t = List<dynamic>.from(v[jsonyVarName] as List);
+      t.forEach((element) {
+        value.add(List<String>.from(element as List));
+      });
+    }
+    else{ _match=false; }
+  }
+}
+
+
+
+
+class JsonyListJsony extends JsonyVar{
+
+  late List<Jsony> elements = [];
+  // constructor wrapper for element objects
+  late Function constructor;
+
+  /// [jsonyVarName] Object name
+  /// [constructor] Wrapper function that returns array element
+  /// [elements] Array objects
+  JsonyListJsony(super.jsonyVarName, Jsony this.constructor(int index), [List<Jsony>? elements]): elements = elements??[];
+  List<Jsony> call() => elements;
+  Map<String,List<dynamic>> get obj{
+    List<dynamic> result = [];
+    elements.forEach((element) {
+      result.add(element.obj);
+    });
+    return {jsonyVarName: result};
+  }
+  void updateElements(dynamic v){
+    try{
+      List<dynamic> _elements = v[jsonyVarName];
+      _elements.forEach((e) {
+        // from json data
+        Map<String, dynamic> element = Map<String, dynamic>.from(e);
+        // create object
+        Jsony newElement = constructor(_elements.indexOf(e));
+        // mix newElement key and data values into new variable
+        Map<String, dynamic> namedElement = {newElement.jsonyVarName: element.values.first};
+        // update object
+        newElement.updateElements(namedElement);
+        elements.add(newElement);
+        !elements.last.isMatched ?_match=false :null;
+      });
+    }
+    catch(e){
+      _match = false;
+      print("${e.toString()} at \"$jsonyVarName\"}.");
+    }
   }
 }