+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## [1.1.0] - 2024/08/03
+
+### Added
+
+- [JsonyListListString] Type
+- [JsonyListJsony] Type. Allows represent json object arrays as iterable Jsony objects.     
+
+### Changed
+
+- Jsony *name* variable name changed to [jsonyVarName]
+- Jsony *update* member name changed to [updateElements]
+- JsonyStringList > [JsonyListString] Name cahenged.
+- JsonyDynamicList > [JsonyListDynamic] Name changed.
+
+### Fixed
+
+
+
 ## 1.0.0
 
 - Initial version.